pyproject.toml
[project]
name = "fastapi-file-upload"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "boto3>=1.42.56",
    "fastapi[standard]>=0.133.0",
    "motor>=3.7.1",
    "pydantic-settings>=2.13.1",
    "pymongo[srv]>=4.16.0",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
    "pre-commit>=4.1.0",
]
